backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
alembic>=1.14.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.12
httpx>=0.28.0
python-dateutil>=2.9.0
# Phase III - AI Chatbot
cohere>=5.0.0
